requirements.txt
discord.py
pandas
requests
openpyxl